Moved source around updated game object methods
