+++ src/game_renderer/Shape.cpp
