+++ src/game_renderer/Shape.h
